backend/requirements.txt
# ===================================
# HEALTHCARE AI - Backend Requirements
# Python 3.9+
# ===================================

# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-Limiter==3.5.0
python-socketio==5.10.0
eventlet==0.33.3
Werkzeug==3.0.1

# Database
pymongo==4.6.1
motor==3.3.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Google AI (Gemini)
google-generativeai==0.3.2
google-api-python-client==2.108.0

# Computer Vision & Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# YOLO (Object Detection)
ultralytics==8.0.228
torch==2.1.1
torchvision==0.16.1

# PDF Generation
reportlab==4.0.7
pypdf2==3.0.1

# Email Services
sendgrid==6.11.0
mailgun2==1.0.5
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.2
email-validator==2.1.0

# Task Scheduling
APScheduler==3.10.4

# Utils
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.1.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
python-json-logger==2.0.7

# CORS & Security Headers
flask-talisman==1.1.0

# Rate Limiting
limits==3.7.0

# Optional: Redis (for caching & rate limiting)
# redis==5.0.1
# flask-redis==0.4.0
